requirements.txt
adium-theme-ubuntu==0.3.4
amqp==2.2.2
billiard==3.5.0.3
boto3==1.4.8
botocore==1.8.8
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
cplex==12.6.3.0
cx-Oracle==6.0.3
decorator==4.0.6
dnspython==1.12.0
docutils==0.14
Flask==0.12.2
Flask-Cors==3.0.3
futures==3.2.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kombu==4.1.0
MarkupSafe==1.0
Mosek==8.1.30
numpy==1.13.3
pandas==0.21.0
patsy==0.4.1
Pillow==3.1.2
protobuf==3.5.0.post1
psutil==5.4.1
pycrypto==2.6.1
pymongo==3.5.1
pymssql==2.1.3
PyMySQL==0.7.11
python-dateutil==2.6.1
pytz==2017.3
redis==2.10.6
requests==2.18.4
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
SQLAlchemy==1.1.15
statsmodels==0.8.0
unity-lens-photos==1.0
urllib3==1.22
vine==1.1.4
websocket-client==0.46.0
Werkzeug==0.12.2
